# mechanicAPI/app/static/swagger.yaml
swagger: '2.0'
info:
  title: "Mechanic Shop API"
  description: "This is a API to manage a mechanic's shop"
  version: "1.0.0"
host: "127.0.0.1:5000"
schemes:
  - "http"
consumes:
  - "application/json"
produces:
  - "application/json"
securityDefinitions:
  bearerAuth:
    type: apiKey
    name: Authorization
    in: header

#* Mechanics Endpoints
paths: 
...
  /part-descriptions/{id}:
    post:
        tags:
          - "Part Descriptions"
        summary: "Endpoint to create a new part description"
        description: "Send a JSON payload including part description information to create a new part description in our database, email must be unique."
        security: []
        parameters:
          - in: "body"
            name: "body"
            description: "Information to create a new part description"
            required: true
            schema:
              $ref: "#/definitions/CreatePart_DescriptionPayload"
        responses:
          200:
            description: "Created Part Description Successfully"
            schema:
              $ref: "#/definitions/AllPart_Descriptions"
            examples:
              application/json:
                brand: "Flowmaster"
                id: 1
                part name: "After Market Exhaust"
                price: "599.99"

    get:
        tags:
          - "Part Descriptions"
        summary: "Returns all Part Descriptions"
        description: "Endpoint to retrieve an array of all part descriptions in the system"
        security: []
        parameters:
          - in: query
            name: page
            type: integer
            description: "Page number for pagination (optional)"
          - in: query
            name: per_page
            type: integer
            default: 10
            description: Items perpage (default 10)
        responses:
          200:
            description: "Successful reterival of Part Descriptions"
            schema:
              $ref: "#/definitions/AllPart_Descriptions"
            examples:
              application/json:
                brand: "Flowmaster"
                id: 1
                part name: "After Market Exhaust"
                price: "599.99"
    put:
        tags:
          - "Part Descriptions"
        summary: "Endpoit to update a part description"
        description: "Endpoint to update a part description."
        security:
          - bearerAuth: []
        parameters:
          - in: 'body'
            name: 'body'
            description: "information required to update a part description"
            required: true
            schema:
              $ref: "#/definitions/UpdatePart_DescriptionPayload"
        responses:
          200:
            description: "Successfully updated part description"
            schema:
              $ref: "#/definitions/UpdatePart_DescriptionResponse"
            examples:
              application/json:
                brand: "Flowmaster"
                id: 1
                part name: "After Market Exhaust"
                price: "599.99"

    delete:
        tags:
          - "Part Descriptions"
        summary: "Endpoint to delete a part description"
        description: "Endpoint to delete a part description."
        security:
          - bearerAuth: []
        responses:
          200:
            description: "Part Description Successfully Deleted"
            schema:
              $ref: "#/definitions/DeletePart_DescriptionResponse"
            examples:
              application/json:
                message: "Successfully deleted part description"

#* Serialized Parts Endpoints
  /serialized-parts/{id}:
    post:
        tags:
          - "Serialized Parts"
        summary: "Endpoint to create a new serialized part"
        description: "Send a JSON payload including serialized part information to create a new serialized part in our database, email must be unique"
        security: []
        parameters:
          - in: "body"
            name: "body"
            description: "Information to create a new serialized part"
            required: true
            schema:
              $ref: "#/definitions/CreateSerialized_PartPayload"
        responses:
          200:
            description: "Created Serialized Part Successfully"
            schema:
              $ref: "#/definitions/AllSerialized_Parts"
            examples:
              application/json:
                desc id: 1
                id: 1
                ticket id: 1
    get:
        tags:
          - "Serialized Parts"
        summary: "Returns all Serialized Parts"
        description: "Endpoint to retrieve an array of all serialized parts in the system"
        security: []
        parameters:
          - in: query
            name: page
            type: integer
            description: "Page number for pagination (optional)"
          - in: query
            name: per_page
            type: integer
            default: 10
            description: Items perpage (default 10)
        responses:
          200:
            description: "Successful reterival of serialized parts"
            schema:
              $ref: "#/definitions/AllSerialized_Parts"
            examples:
              application/json:
                desc id: 1
                id: 1
                ticket id: 1
    put:
        tags:
          - "Serialized Parts"
        summary: "Endpoit to update a serialized part"
        description: "Endpoint to update a serialized part."
        security:
          - bearerAuth: []
        parameters:
          - in: 'body'
            name: 'body'
            description: "information required to update a serialized part"
            required: true
            schema:
              $ref: "#/definitions/UpdateSerialized_PartPayload"
        responses:
          200:
            description: "Successfully updated serialized part"
            schema:
              $ref: "#/definitions/UpdateSerialized_PartResponse"
            examples:
              application/json:
                desc id: 1
                id: 1
                ticket id: 1
    delete:
        tags:
          - "Serialized Parts"
        summary: "Endpoint to delete a serialized part"
        description: "Endpoint to delete a serialized part."
        security:
          - bearerAuth: []
        responses:
          200:
            description: "Serialized Part Successfully Deleted"
            schema:
              $ref: "#/definitions/DeleteSerialized_PartResponse"
            examples:
              application/json:
                message: "Successfully deleted serialized part"

#* Service Tickets Endpoints
...
definitions: 
...
  CreatePart_DescriptionPayload:
    type: "object"
    properties:
      part_name:
        type: "string"
      brand:
        type: "string"
      price:
        type: "string"
    required:
      - part name
      - brand
      - price
...
  AllPart_Descriptions:
    type: "array"
    properties:
      brand:
        type: "string"
      id:
        type: "integer"
      part_name:
        type: "string"
      price:
        type: "string"
  UpdatePart_DescriptionPayload:
    type: "object"
    properties:
      brand:
        type: "string"
      id:
        type: "integer"
      part_name:
        type: "string"
      price:
        type: "string"

    required:
      - brand
      - id
      - part name
      - price
  UpdatePart_DescriptionResponse:
    type: 'object'
    properties:
      brand:
        type: "string"
      id:
        type: "integer"
      part_name:
        type: "string"
      price:
        type: "string"
...
  CreateSerialized_PartPayload:
    type: "object"
    properties:
      desc_id:
        type: "integer"
      ticket_id:
        type: "integer"
    required:
      - desc id
      - ticket id
...
  AllSerialized_Parts:
    type: "array"
    properties:
      id:
        type: "integer"
      desc_id:
        type: "integer"
      ticket_id:
        type: "integer"
  UpdateSerialized_PartPayload:
    type: "object"
    properties:
      desc_id:
        type: "integer"
      ticket_id:
        type: "integer"

    required:
      - desc id
      - ticket id
  UpdateSerialized_PartResponse:
    type: 'object'
    properties:
      id:
        type: "integer"
      desc_id:
        type: "integer"
      ticket_id:
        type: "integer"
...